requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
aiohttp==3.9.1
networkx==3.2.1
pydantic==2.5.0
python-dotenv==1.0.0
typing-extensions==4.8.0
jinja2==3.1.2
